Refactor to fix pylint issues
