+++ pylintrc
@@ -77,7 +77,10 @@ confidence=
 disable=missing-module-docstring,
         missing-class-docstring,
         too-many-instance-attributes,
-        missing-function-docstring
+        missing-function-docstring,
+        logging-fstring-interpolation,
+        too-many-arguments,
+        too-many-locals
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
@@ -206,7 +209,7 @@ ignored-classes=optparse.Values,thread._local,_thread._local
 # (useful for modules/projects where namespaces are manipulated during runtime
 # and thus existing member attributes cannot be deduced by static analysis). It
 # supports qualified module names, as well as Unix pattern matching.
-ignored-modules=
+ignored-modules=trogon,sqlglot,tqdm,pydantic
 
 # Show a hint with possible names when a member name was not found. The aspect
 # of finding the hint is based on edit distance.
