+++ src/hdx_cli/cli_interface/common/cached_operations.py
@@ -1,10 +1,5 @@
-import json
-
-import requests
-
 from ...library_api.common.config_constants import HDX_CONFIG_DIR
 from ...library_api.common.context import ProfileUserContext
-from ...library_api.common.exceptions import HdxCliException, ResourceNotFoundException
 from ...library_api.common.generic_resource import access_resource
 from ...library_api.common.logging import get_logger
 from ...library_api.utility.decorators import find_in_disk_cache
