+++ src/hdx_cli/library_api/ddl/common_intermediate_representation.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from enum import Enum
 from typing import Any, Callable, Dict, List, Optional, Set, Union
 
 __all__ = [
