+++ src/hdx_cli/library_api/ddl/interfaces.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
+from typing import Any, Dict, Iterator, List, Tuple, Union
 
 from .common_intermediate_representation import (
     ColumnDefinition,
